# editor/server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "editor API",
    "version": "0.1.0",
    "description": "Jinja2 テンプレート GUI 編集アプリのバックエンド API。\n\n`@editor/shared` の Repository 契約(Template / Auth / User / History / Part)を REST として設計したもの。一部は Phase 2 の設計(ハンドラ未実装)。"
  },
  "servers": [
    {
      "url": "/api",
      "description": "同一オリジン (server が SPA も配信)"
    }
  ],
...
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/auth/init-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "パスワード変更(本人・現行パスワードによる所有証明が必要)",
        "operationId": "initPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordInitRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "変更完了(他端末のセッションは失効)"
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "403": {
            "description": "本人以外を指定した、またはレート制限で拒否 (kind=forbidden)。後者の `code` は login と同じ LOGIN_RATE_LIMITED / LOGIN_BUSY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
...
    "/templates/{templateId}/notes": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "パーツ単位コメントを取得(版インスタンスごとに独立。親投稿と返信が混在)",
        "operationId": "listNotes",
        "parameters": [
          {
            "in": "path",
            "name": "templateId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "投稿の配列(作成日時の昇順。親投稿と返信が混在し、返信は replyTo で親を指す)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PartNoteEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "404": {
            "description": "対象が存在しない (kind=not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "notes"
        ],
        "summary": "パーツ単位コメントへ投稿を追加",
        "operationId": "addNote",
        "parameters": [
          {
            "in": "path",
            "name": "templateId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "追加した投稿",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartNoteEntry"
                }
              }
            }
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{templateId}/notes/{entryId}": {
      "patch": {
        "tags": [
          "notes"
        ],
        "summary": "投稿の本文・状態を更新",
        "operationId": "updateNote",
        "parameters": [
          {
            "in": "path",
            "name": "templateId",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "entryId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新後の投稿(状態の変更は返信にも伝播する)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartNoteEntry"
                }
              }
            }
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notes"
        ],
        "summary": "コメントの投稿を削除",
        "operationId": "deleteNote",
        "parameters": [
          {
            "in": "path",
            "name": "templateId",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "entryId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "削除完了"
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PasswordInitRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "currentPassword": {
            "type": "string",
            "description": "現在のパスワード(所有証明)"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "currentPassword",
          "newPassword"
        ]
      },
...
      "AddNoteRequest": {
        "type": "object",
        "properties": {
          "pathKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "description": "パーツ構造キー(文書全体でのアンカー#通し番号)"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 65536,
            "description": "投稿本文"
          },
          "replyTo": {
            "default": null,
            "description": "返信先の親投稿 ID。null なら親投稿として追加する",
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "pathKey",
          "content"
        ]
      },
      "UpdateNoteRequest": {
        "type": "object",
        "properties": {
          "content": {
            "description": "投稿本文",
            "type": "string",
            "minLength": 1,
            "maxLength": 65536
          },
          "status": {
            "$ref": "#/components/schemas/NoteStatus"
          }
        }
      },
      "NoteStatus": {
        "type": "string",
        "enum": [
          "open",
          "resolved"
        ]
      },
...
      "AppError": {
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/AppErrorKind"
          },
          "message": {
            "type": "string",
            "description": "ユーザ向け(JP)。常に表示して安全"
          },
          "code": {
            "description": "機械可読コード 例: 'USER_DISABLED'",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "message"
        ],
        "additionalProperties": false
      },
      "AppErrorKind": {
        "type": "string",
        "enum": [
          "not_found",
          "validation",
          "unauthorized",
          "forbidden",
          "conflict",
          "network",
          "unexpected"
        ]
      },
...
      "PartNoteEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "投稿 ID(UUID)"
          },
          "templateId": {
            "type": "string",
            "description": "投稿が属する版インスタンス ID"
          },
          "pathKey": {
            "type": "string",
            "description": "パーツ構造キー(文書全体でのアンカー#通し番号)"
          },
          "content": {
            "type": "string",
            "description": "投稿本文"
          },
          "createdAt": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "本文が編集された場合のみ"
          },
          "updatedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "description": "未対応 / 解決済み。返信は親と同じ値",
            "$ref": "#/components/schemas/NoteStatus"
          },
          "replyTo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "親投稿の ID。null なら親(スレッドの起点)"
          }
        },
        "required": [
          "id",
          "templateId",
          "pathKey",
          "content",
          "createdAt",
          "createdBy",
          "updatedAt",
          "updatedBy",
          "status",
          "replyTo"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "editor.sid",
        "description": "ログイン時に発行されるセッション cookie。"
      }
    }
  }
}